==> Notebooks/1. Data Cleaning.ipynb <==
# Data Cleaning

## Data Collection 

### Importing the necessary libraries and personal drive

# Importing libraries
from google.colab import drive
import pandas as pd 
import matplotlib.pyplot as plt
import numpy as np

# Mount Google Drive to this Notebook instance.
drive.mount('/content/drive')

## Loading the raw data from the IATI website
The link to the website: https://iatistandard.org/en/

# Creating a dataframe with the original dataset
file = '/content/drive/MyDrive/Data/IATI_raw_data.csv'
df = pd.read_csv(file)

## Briefly explore the data 

# Viewing the first rows if the raw dataset
df.head()

# View all columns
df.columns

# View the data information
df.info()

## Cleaning and formattung the data 

#Dropping the unnecessary columns 
df = df.drop(columns = ['activity_status_code', 'activity_scope_code', 'capital_spend_percentage', 'title_xml',
                           'reporting_org_xml', 'reporting_org_ref', 'reporting_org_type_code', 'reporting_org_type_name',
                           'reporting_org_secondary_reporter', 'reporting_org_narrative', 'description_xml', 
                           'participating_org_xml', 'activity_date_xml', 'activity_date_type', 
                           'recipient_country_xml', 'sector_xml', 'location_xml', 'budget_xml', 
                           'recipient_region_xml', 'activity_date_end_actual_f', 
                         'description_type', 'participating_org_ref', 'participating_org_type', 'participating_org_role',
                       'activity_date_iso_date', 'activity_date_end_planned_f', 'activity_date_start_actual_f', 'activity_date_start_planned_f',
                       'recipient_country_percentage', 'sector_vocabulary', 'sector_vocabulary_uri', 'sector_percentage',
                       'location_point_pos', 'location_name_narrative_text', 'budget_type', 'budget_status', 'budget_period_start_iso_date',
                        'budget_period_end_iso_date', 'budget_value_date', 'budget_value_usd_sum', 'budget_imf_link', 'budget_usd_conversion_rate',
                        'location_exactness_code', 'location_class_code', 'recipient_region_code', 'recipient_region_vocabulary', 'recipient_region_vocabulary_uri',
                        'activity_date_narrative_lang', 'sector_narrative_lang', 'sector_narrative_text', 'activity_date_narrative_text', 'recipient_country_narrative_text',
                        'recipient_region_narrative_text', 'recipient_region_percentage', 'location_reach_code', 'location_id_code', 'location_id_vocabulary', 
                        'location_administrative_level', 'location_feature_designation_code', 'location_administrative_vocabulary', 'location_description_narrative_text',
                        'location_activity_description_narrative_text', 'location_administrative_code',
                          'description_narrative_text', 'participating_org_narrative_text', 'activity_date_start_planned',
                  'budget_value_currency', 'budget_value', 'activity_date_narrative', 'activity_date_end_actual', 
                  'location_description_narrative_lang', 'location_description_narrative', 'location_activity_description_narrative_lang', 
                  'location_activity_description_narrative', 'recipient_country_narrative', 'recipient_region_narrative_lang',
                        'default_currency', 'title_narrative_lang', 'participating_org_narrative',
                            'participating_org_narrative_lang', 'activity_date_end_planned', 'activity_date_start_actual',
                            'recipient_country_code', 'sector_code', 'location_ref', 'location_name_narrative', 'location_name_narrative_lang',
                            'budget_value_usd', 'recipient_region_name', 'sector_narrative', 'recipient_country_narrative_lang', 
                            'recipient_region_narrative', 'default_currency', 'iati_identifier', 'recipient_country_name', 'title_narrative_lang'
                         ] )

# Viewing the columns that are left 
df.columns

# Viewing the first rows of the 
df.head

# Keeping the descriptions in english only
df = df[df['description_lang'] == 'en']

# Resetting the index
df = df.reset_index()

# Dropping the new index column
df = df.drop('index', axis=1)

# Deleting all the rows with no descriptions 
df = df[df['description_narrative'].notnull()]

# We need to count the words in each description to keep only the longest 

# Creating a list with all the descriptions 
sentences = [i for i in df['description_narrative']]

# Setting an index list 
index = [i for i in range(len(sentences))] 

# Creating an empty list for the number of words which will later be fed 
number_words = []

#looping in the indexes
for i in index:
    #split the words in every sentence
    word = len(sentences[i].split())
    #appending the number of words in the list number_words
    number_words.append(word)

# Adding the list as a column to the main dataframe 
df['number_words'] = number_words

# View the dataset 
df.head()

# See the main statistics of the descriptions 
df.describe()

# We keep the longest descriptions
df = df[df['number_words'] > 207]

# Randomly selecting 5,000 descriptions of the dataframe
df = df.sample(n = 5000)

# create excel writer object
writer = pd.ExcelWriter('documents.xlsx')
# write dataframe to excel
df.to_excel(writer)
# save the excel
writer.save()
print('DataFrame is written successfully to Excel File.')

## Once manually annotated, we re-load the data 'annotated_documents' for last cleaning steps


df = pd.read_excel('/content/drive/MyDrive/Data/annotated_documents.xlsx')
df

# This was done after the manual annotating of the gold summaries
# We load the new dataframe
df = pd.read_excel('/content/drive/MyDrive/Data/annotated_documents.xlsx')

# Remove unwanted characters from both the documents and the summaries
df['description_narrative'].replace("\n", "")

df['description_narrative'] = df['description_narrative'].str.replace("\n", "")

df['TLDR'] = df['TLDR'].str.replace("\\", "")

df['TLDR'].replace("\n", "")

df['TLDR'] = df['TLDR'].str.replace("\n", "")

df['TLDR'] = df['TLDR'].str.replace("\\", "")

# Remove unnecessary added Unnamed: 0 column and title_narrative, which was helpful to before create the gold summaries manually
df = df.drop(columns=['Unnamed: 0', 'title_narrative', 'description_lang', 'number_words'])

# Renaming columns 
df.columns = ['document_text', 'summary_text']

# Dropping rows that were not desired
df = df.drop([31,45,480,518,916,
              1340,1372,1492,1562,
              2167, 2345,2346,2613, 
              2752, 2824])

df.head()

# Exporting as final_clean in a csv format for the next step of data wrangling
df.to_csv('SummaDevDocs_raw.csv', index=False)
