==> Notebooks/2. Data Wrangling .ipynb <==
# Dissertation

# Data Wrangling

## Importing and installing the necessary libraries

!pip install sentence_transformers

!python -m spacy download en_core_web_lg
# restart session

# Importing necessary librairies
from google.colab import drive
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import pickle 
from sentence_transformers import SentenceTransformer
from sklearn.metrics.pairwise import cosine_similarity
import spacy

## Importing Personal Drive to access the data

# Mount Google Drive to this Notebook instance.
drive.mount('/content/drive')

## Splitting each description and TLDR into list of sentences 

The descriptions and TLDRs are read into the spacy tokenizer format using the large English mode.
These are then split into sentences using spaCy's .sents method. The spacy tokenizer format carries a lot of memory overhead and the list is converted to a list of strings before completing this step.

BERT is sophisticated enough to parse meaning from all the nuances of language and steps such as stop word removal, stemming and lower-case transformations are purposely ignored. 


# Loading the clean dataset intp a pandas dataframe
file = '/content/drive/MyDrive/Data/SummaDevDocs_raw.csv'
df = pd.read_csv(file)

# Using spaCy to split text into sentences, clean up short sentences and embed using BERT via the sentence-transformer package
### Helper function

def text_to_sent_list(text, 
                      nlp = spacy.load("en_core_web_lg"), 
                      embedder = SentenceTransformer('distilbert-base-nli-mean-tokens'),
                      min_len=0):
  
    ''' Returns cleaned article sentences and BERT sentence embeddings'''
    
    #convert to list of sentences
    text = nlp(text)
    sents = list(text.sents)
    #remove short sentences by threshhold                                                                                                
    sents_clean = [sentence.text for sentence in sents if len(sentence)> min_len]
    #remove entries with empty list
    sents_clean = [sentence for sentence in sents_clean if len(sentence)!=0]
    #embed sentences (deafult uses BERT SentenceTransformer)
    sents_embedding= np.array(embedder.encode(sents_clean, convert_to_tensor=True))
    
    return sents_clean, sents_embedding

#load nlp and embedder
nlp = spacy.load("en_core_web_lg")
embedder = SentenceTransformer('distilbert-base-nli-mean-tokens')

#extract clean sentence list and sentence embedding for each article TEXT
f = lambda text: text_to_sent_list(text, nlp=nlp, embedder=embedder, min_len=0)
s_interim_tuple = df['document_text'].apply(f)

df['text_clean'] = s_interim_tuple.apply(lambda x: x[0])
df['text_embedding'] = s_interim_tuple.apply(lambda x: x[1])

#extract clean sentence list and sentence embedding for each article SUMMARY
f = lambda summ: text_to_sent_list(summ, nlp=nlp, embedder=embedder, min_len=0)
s_interim_tuple = df['summary_text'].apply(f)

df['summary_clean'] = s_interim_tuple.apply(lambda x: x[0])
df['summary_embedding'] = s_interim_tuple.apply(lambda x: x[1])

A cosine similarity is calculated for each summary sentence shere each entry represents this measure between each article sentence with that summary sentence. SK-learn's inbuilt cosine similarity function is used. 

# Calculate the cosine-similarity matrix between each summary sentence and the article sentences
#helper function
def find_sim_single_summary(summary_sentence_embed, doc_emedding):
    '''returns array of indices for max cosine sim per summary sentences'''
    cos_sim_mat = cosine_similarity(doc_emedding, summary_sentence_embed)
    idx_arr = np.argmax(cos_sim_mat, axis=0)
    
    return idx_arr

#main function
def label_sent_in_summary(s_text, s_summary):
    '''returns index list and binary target labels in an array'''
    doc_num = s_text.shape[0]
    #initialize zeros
    labels = [np.zeros(doc.shape[0]) for doc in s_text.tolist()] 
    #calc idx for most similar
    idx_list = [np.sort(find_sim_single_summary(s_summary[j], s_text[j])) for j 
                                                            in range(doc_num)]
    for j in range(doc_num):
        labels[j][idx_list[j]]= 1 
        
    return idx_list, labels

#get index list and target labels
idx_list, labels = label_sent_in_summary(df.text_embedding, df.summary_embedding)

#wrap in dataframe
df['labels'] = labels

# View the data
df

## Adding doc_label manually

# Creating a list of numbers from 0 to 2985, the number of documents 
document_number = list(range(0, 2985))

# Computing the number of sentences in every single documents
#creatin an empty list 
num_sentences = []

#looping through every list in labels
for i in df['labels']:
    #appending the lenght of each list to the num_sentences lists 
    num_sentences.append(len(i))

# Creating an empty list of labels to later assign each sentence the label of the document it belongs to
doc_label = []

# Looping through each number from O to 2985, and creating a list of lists
# Each list within the list will correspond to the label of the document times the number of sentences in the document 
# This allows to track back each sentence to its original document
for i in document_number:
    labels = [i]*num_sentences[i]
    doc_label.append(labels)
    
# Adding the document labels as a column in the main dataframe 
df['doc_label']=doc_label

# View the data with the updated column 
df

# Exporting the file as a pickle
output_file = 'SummaDevDocs_preprocesssed.pickle'

with open(output_file, 'wb') as handle:                                     
    pickle.dump(df, handle)
