==> Notebooks/4. Training and Test Data Set Split.ipynb <==
# 1. Setup

## 1.1.Importing and installing the necessary libraries


# Importing necessary librairies
from google.colab import drive
import numpy as np
import pandas as pd
import pickle 
from sklearn.model_selection import train_test_split

## 1.2. Importing personal drive to file in order to later load the data and the saved fine tuned model


# Mount Google Drive to this Notebook instance.
drive.mount('/content/drive')

## 2. Load and Format data

## 2.1. Loading the clean data




# We load the cleaned data from a pickle format
file ='/content/drive/MyDrive/Data/SummaDevDocs_preprocessed.pickle'
df = pd.read_pickle(file)

# Looking at the data
df

## 2.2. Creating the train and test sets in the right format

### 2.2.1. Creating the train set, with each row being a sentence 



# Splitting the dataframe into a train and a test set, keeping 20% for the test set. 
train_dataset, test_dataset = train_test_split(df, test_size=0.2)

# We wish to split each sentence individually for the purpose of classification. 
#for this reason, we create three empty lists to later feed them with the individual sentences, associated with their label and document label 
sentences, labels, document_labels = [], [], []

# We do the below to feed the empty lists
#looping through the embeddings, the document labels, the text and the target labels
for doc_x, doc_y, doc_label in zip(train_dataset['text_clean'], 
                                   train_dataset['labels'], 
                                   train_dataset['doc_label']):
                                          
#looping through each item within the text, the target labels, the document labels
    for i, (sent, target, label) in enumerate(zip(doc_x, doc_y, doc_label)):
        
        #appending each item to the empty lists previously created 
        sentences.append(sent) #this will be used to train the data points 
        labels.append(target) #this will be used to train the data points
        document_labels.append(label) #this will be used to retrieve the intial document the sentence belongs to 

# We then create a dataframe out of the created lists 
train_dataset = pd.DataFrame({'sentence': sentences,
                             'label' :  labels,
                             'document_label': document_labels
                             })

# For precision, we convert the current float labels to integers 
#creation of an empty list
new_label = []

#looping in the label column, converting each label to an integer and appending to the new_label list
for i in train_dataset['label']:
  i = int(i)
  new_label.append(i)  

#updating the label column with the new_label list
train_dataset['label'] = new_label

# Exporting to csv file
train_dataset.to_csv('train_dataset.csv', index=False)

# Viewing the pre-formatted train_dataset
train_dataset.head()

### 2.2.2. Creating the test set, with each row being a sentence


# We repeat the same process for the test_dataset

# Creating empty lists 
sentences, labels, document_labels = [], [], [] 

#looping through the embeddings, the document labels, the text and the target labels
for doc_x, doc_y, doc_label in zip(test_dataset['text_clean'], 
                                   test_dataset['labels'], 
                                   test_dataset['doc_label']):
                                          
    #looping through each item within the text, the target labels, the document labels
    for i, (sent, target, label) in enumerate(zip(doc_x, doc_y, doc_label )):
        
        #appending each item to the empty lists previously created 
        sentences.append(sent) #this will be used to train the data points 
        labels.append(target) #this will be used to train the data points
        document_labels.append(label) #this will be used to retrieve the intial document the sentence belongs to 

# Creating a new test_dataset dataframe with appropriate lists 
test_dataset = pd.DataFrame({'sentence': sentences,
                             'label' :  labels,
                             'document_label': document_labels,
                             })

# Converting float labels to integers 
new_label = []

for i in test_dataset['label']:
  i = int(i)
  new_label.append(i)   

#updating the label column with the new_label list
test_dataset['label'] = new_label

# Exporting to csv file
test_dataset.to_csv('test_dataset.csv', index=False)

# Viewing the pre-formatted train_dataset
test_dataset.head()
